bring back cinic downloading
