+++ utils/augmentors.py
@@ -1,8 +1,8 @@
 import torch
 import numpy as np
+import torchvision.transforms as transforms
 
-
-class Cutout(object):
+class Cutout:
     """
     Taken primarily from: https://github.com/uoguelph-mlrg/Cutout
     Randomly mask out one or more patches from an image.
@@ -41,3 +41,21 @@ class Cutout(object):
         img = img * mask
 
         return img
+
+class SimCLRTransform:
+    """
+    credit: https://github.com/sthalles/SimCLR/
+    """
+    def __init__(self, size, s=1, n_views=2):
+        color_jitter = transforms.ColorJitter(0.8 * s, 0.8 * s, 0.8 * s, 0.2 * s)
+        base_transform = transforms.Compose([transforms.RandomResizedCrop(size=size),
+                                              transforms.RandomHorizontalFlip(),
+                                              transforms.RandomApply([color_jitter], p=0.8),
+                                              transforms.RandomGrayscale(p=0.2),
+                                              transforms.GaussianBlur(kernel_size=int(0.1 * size)),
+                                              transforms.ToTensor()])
+        self.base_transform = base_transform
+        self.n_views = n_views
+
+    def __call__(self, x):
+        return [self.base_transform(x) for i in range(self.n_views)]
